Adds support for AR Orientation Tracking (#198)
* App Store 1024x1024 Icon

This eliminates an xcode compile warning

* Fix CLHeading.headingAccuracy variable definition

Was CLLocationDegrees, now CLLocationDirection.  Worth noting that both are doubles so the code compiled w/o error, but it's now correct.

* Camera tracking mechanism sync to device orientation better, especially after pause/resume

Was ARWorldTracdkingConfiguration, now AROrientationTrackingConfiguration.  Heading (Euler.y) now tracks soley to AHRS, before it tried to sense and track to planes - which doesn't make sense for an outside environment.

* Added a configuration option for World Tracking versus Orientation Tracking.  The default is World Tracking.

* Updated the changelog.

* Added more (defaulted) parameters to the convenience initializer and re-commented it.  Also addressed an issue Marc found in the PR.